refactor(frontend): Clean up graph import & export logic
